requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
discord.py==2.5.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
orjson==3.11.2
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
redis==6.4.0
sniffio==1.3.1
typing_extensions==4.14.1
yarl==1.20.1
google-genai==1.29.0
